EpicEvents/requirements.txt
asgiref==3.5.2
Django==4.0.4
django-extensions==3.1.5
django-filter==21.1
django-rename-app==0.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
drf-writable-nested==0.6.3
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
